Add activating constructor for string

+++ src/json/json_string.h
@@ -183,7 +183,7 @@ namespace format
        * otherwise return string as negative value.
        * @return Number of characters read
        */
-      long /** TODO: long */
+      long long
       __string (wchar_t & endc) const noexcept;
 
       /**
